refactor: migrated tests back in checkoutUI file
